feature: add customtab helper and navigation to support external url
