+++ app/src/main/kotlin/com/wire/android/ui/WireActivity.kt
@@ -17,10 +17,10 @@ import com.wire.android.navigation.NavigationManager
 import com.wire.android.navigation.navigateToItem
 import com.wire.android.ui.theme.WireTheme
 import dagger.hilt.android.AndroidEntryPoint
+import javax.inject.Inject
 import kotlinx.coroutines.CoroutineScope
 import kotlinx.coroutines.flow.launchIn
 import kotlinx.coroutines.flow.onEach
-import javax.inject.Inject
 
 @ExperimentalMaterial3Api
 @ExperimentalAnimationApi
@@ -66,9 +66,6 @@ class WireActivity : AppCompatActivity() {
             navigationManager.navigateBack
                 .onEach { navController.popBackStack() }
                 .launchIn(scope)
-
         }
     }
-
 }
-
